test(stella-cli): split config/tests.rs into cohesive files under config/tests/
config/tests.rs held 41 tests in 1462 lines, 38 lines from the strict
1500-line ceiling. Group by subject into config/tests/{parity,credential,
resolution,docs_sync}.rs, following the config/tests/trusted_engine.rs
pattern already started here: parity.rs holds the provider_parity matrix
completeness checks, credential.rs the key/credential-source resolution
tests, resolution.rs the broader Config::load_with_settings scenarios
(plus its own reload_fixture helper), and docs_sync.rs the
provider-cards.tsx drift check. The settings_from helper stays directly
in config/tests.rs since both credential.rs and resolution.rs call it
through the same super::* accessibility trusted_engine.rs already relies
on. Pure test move, no production change.

Closes #3701
